requirements.txt
# requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
requests>=2.31.0
streamlit>=1.31.1
pandas>=2.0.3
xlsxwriter>=3.1.9